refactor: di를 app 모듈에서 data 모듈로 이동
