+++ app/build.gradle
@@ -5,9 +5,6 @@ plugins {
     id 'kotlin-kapt'
 }
 
-def localProperties = new Properties()
-localProperties.load(new FileInputStream(rootProject.file("local.properties")))
-
 android {
     namespace 'com.example.interviewmate'
     compileSdk 33
@@ -20,8 +17,6 @@ android {
         versionName "1.0"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-
-        buildConfigField "String", "INTERVIEWMATE_URL", localProperties['interviewmate_url']
     }
 
     buildTypes {
@@ -54,17 +49,6 @@ dependencies {
     //Hilt
     implementation 'com.google.dagger:hilt-android:2.44'
     kapt 'com.google.dagger:hilt-compiler:2.44'
-
-    // Retrofit
-    implementation 'com.squareup.retrofit2:retrofit:2.9.0'
-    implementation 'com.squareup.retrofit2:converter-moshi:2.9.0'
-
-    // OkHttp3
-    implementation 'com.squareup.okhttp3:okhttp:4.10.0'
-    implementation 'com.squareup.okhttp3:logging-interceptor:4.10.0'
-
-    // Moshi
-    implementation 'com.squareup.moshi:moshi-kotlin:1.14.0'
 }
 
 kapt {
